XMLHttpRequest/exception-InvalidAccessError-responseType.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>XMLHttpRequest Test: InvalidAccessError- responseType</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<meta name="assert" content="Check if exception InvalidAccessError should be thrown when async is false and the responseType attribute value is not the empty string">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="log"></div>

<script>

async_test(function (t) {
  var client = new XMLHttpRequest();
  client.open("GET", "resources/headers.py", false);
  t.step(function () {
    assert_throws("InvalidAccessError", function () { client.responseType = "text"; });
  });
  t.done();
}, document.title);

</script>
